P2Y (Iron) tests added

+++ libs/UTILS_upload.pm
@@ -29,15 +29,22 @@ $archive =~ s/\//_/;
 $test_dir   =$upload_test;
 $test_subdir=".";
 if ( scalar @paths > 1) {
-  $test_dir=@paths[0];
-  $test_subdir=$upload_test;
-  $test_subdir =~ s/$test_dir\///;
+ $test_dir=@paths[0];
+ $test_subdir=$upload_test;
+ $test_subdir =~ s/$test_dir\///;
 }
 chdir("$TESTS_folder/$test_dir");
-&command("find $test_subdir -name 'ns.*' -o -name 'ndb*gkkp*' -o -name 'ndb*Double*' | grep -v 'ROBOT_'| xargs tar cvf $archive.tar");
-&command("gzip $archive.tar");
-&FTP_upload_it("$archive.tar.gz","testing-robots/databases");
-&command("rm -f $test_subdir/SAVE/*");
-&command("git add $test_subdir/SAVE");
+#
+if (-d $test_subdir){
+ &command("find $test_subdir -name 'ns.*' -o -name 'ndb*gkkp*' -o -name 'ndb*Double*' | grep -v 'ROBOT_'| xargs tar cvf $archive.tar");
+ &command("find . -type f $test_subdir | grep -v 'ROBOT_'| xargs tar cvf $archive.tar");
+ &command("gzip $archive.tar");
+ &FTP_upload_it("$archive.tar.gz","testing-robots/databases");
+ &command("rm -f $test_subdir/SAVE/*");
+ &command("git add $test_subdir/SAVE");
+}else
+{
+ &FTP_upload_it("$upload_test","testing-robots/databases");
+}
 }
 1;
